Modularizes color linting tests
Refactors color linting logic by splitting tests into separate modules for better maintainability and reusability. Updates references in the main test file to use the new modular structure.
